Some fixes

+++ lesson2/task6_l02.py
@@ -1,3 +1,3 @@
 number = float(input("Q:Enter the int or float:\nA:"))
-print("hex:", hex(int(number)))
-print("oct", oct(int(number)))
+print("hex:", float(hex((number))))
+print("oct", float(oct((number))))
